requirements.txt
boto3
botocore
pypdfium2
faiss-cpu
PyMuPDF
langchain
langchain-community
langchain-openai
langchain-aws
tabula-py
tqdm
streamlit
jpype1
requests
numpy
ipython
openai
python-dotenv
transformers
torch
pillow
scikit-learn